+++ Assets/Tests/JsonConvertersTests.cs
@@ -1,7 +1,6 @@
 using NUnit.Framework;
 using PAC.Json;
 using PAC.Extensions;
-using PAC.KeyboardShortcuts;
 using UnityEngine;
 
 namespace PAC.Tests
@@ -35,7 +34,7 @@ namespace PAC.Tests
             JsonList jsonList = new JsonList(new JsonFloat(0.2f), new JsonFloat(0.4f));
             Vector2 expected = new Vector2(0.2f, 0.4f);
 
-            Assert.AreEqual(JsonConverter.FromJson<Vector2>(jsonList, converters, false), expected);
+            Assert.AreEqual(expected, JsonConverter.FromJson<Vector2>(jsonList, converters, false));
 
             // List too short
             Assert.Catch(() => JsonConverter.FromJson<Vector2>(new JsonList(new JsonFloat(0.2f)), converters, false));
@@ -70,7 +69,7 @@ namespace PAC.Tests
             JsonList jsonList = new JsonList(new JsonFloat(0.2f), new JsonFloat(0.4f), new JsonFloat(0.1567f));
             Vector3 expected = new Vector3(0.2f, 0.4f, 0.1567f);
 
-            Assert.AreEqual(JsonConverter.FromJson<Vector3>(jsonList, converters, false), expected);
+            Assert.AreEqual(expected, JsonConverter.FromJson<Vector3>(jsonList, converters, false));
 
             // List too short
             Assert.Catch(() => JsonConverter.FromJson<Vector3>(new JsonList(new JsonFloat(0.2f), new JsonFloat(0.4f)), converters, false));
@@ -106,7 +105,7 @@ namespace PAC.Tests
             JsonList jsonList = new JsonList(new JsonFloat(0.2f), new JsonFloat(0.4f), new JsonFloat(0.1567f), new JsonFloat(0.95f));
             Color expected = new Color(0.2f, 0.4f, 0.1567f, 0.95f);
 
-            Assert.AreEqual(JsonConverter.FromJson<Color>(jsonList, converters, false), expected);
+            Assert.AreEqual(expected, JsonConverter.FromJson<Color>(jsonList, converters, false));
 
             // List too short
             Assert.Catch(() => JsonConverter.FromJson<Color>(new JsonList(new JsonFloat(0.2f), new JsonFloat(0.4f), new JsonFloat(0.1567f)), converters, false));
@@ -178,110 +177,14 @@ namespace PAC.Tests
 
             Assert.AreEqual(converted.width, expected.width);
             Assert.AreEqual(converted.height, expected.height);
-            Assert.AreEqual(converted.GetPixel(0, 0), expected.GetPixel(0, 0));
-            Assert.AreEqual(converted.GetPixel(1, 0), expected.GetPixel(1, 0));
-            Assert.AreEqual(converted.GetPixel(0, 1), expected.GetPixel(0, 1));
-            Assert.AreEqual(converted.GetPixel(1, 1), expected.GetPixel(1, 1));
+            Assert.AreEqual(expected.GetPixel(0, 0), converted.GetPixel(0, 0));
+            Assert.AreEqual(expected.GetPixel(1, 0), converted.GetPixel(1, 0));
+            Assert.AreEqual(expected.GetPixel(0, 1), converted.GetPixel(0, 1));
+            Assert.AreEqual(expected.GetPixel(1, 1), converted.GetPixel(1, 1));
 
             // Num of pixels != width * height
             jsonObj["width"] = new JsonInt(3);
             Assert.Catch(() => JsonConverter.FromJson<Texture2D>(jsonObj, converters, false));
-        }
-
-        /// <summary>
-        /// Checks that ToJson() works properly for the custom JSON converter for type CustomKeyCode.
-        /// </summary>
-        [Test]
-        [Category("JSON"), Category("Keyboard Shortcuts")]
-        public void CustomKeyCodeToJson()
-        {
-            JsonConverterSet converters = new JsonConverterSet(new CustomKeyCodeJsonConverter());
-
-            (CustomKeyCode, JsonData)[] testCases =
-            {
-                (CustomKeyCode.Ctrl, new JsonString("Ctrl")),
-                (CustomKeyCode._2, new JsonString("2")),
-                (CustomKeyCode.Plus, new JsonString("+")),
-                (CustomKeyCode.Shift, new JsonString("Shift")),
-                (CustomKeyCode.GreaterThan, new JsonString(">")),
-            };
-
-            foreach ((CustomKeyCode keyCode, JsonData expected) in testCases)
-            {
-                Assert.True(JsonData.HaveSameData(JsonConverter.ToJson(keyCode, converters, false), expected));
-            }
-        }
-
-        /// <summary>
-        /// Checks that FromJson() works properly for the custom JSON converter for type CustomKeyCode.
-        /// </summary>
-        [Test]
-        [Category("JSON"), Category("Keyboard Shortcuts")]
-        public void CustomKeyCodeFromJson()
-        {
-            JsonConverterSet converters = new JsonConverterSet(new CustomKeyCodeJsonConverter());
-
-            (CustomKeyCode, JsonData)[] testCases =
-            {
-                (CustomKeyCode.Ctrl, new JsonString("Ctrl")),
-                (CustomKeyCode._2, new JsonString("2")),
-                (CustomKeyCode.Plus, new JsonString("+")),
-                (CustomKeyCode.Shift, new JsonString("Shift")),
-                (CustomKeyCode.GreaterThan, new JsonString(">")),
-            };
-
-            foreach ((CustomKeyCode expected, JsonData jsonData) in testCases)
-            {
-                Assert.AreEqual(JsonConverter.FromJson<CustomKeyCode>(jsonData, converters, false), expected);
-            }
-        }
-
-        /// <summary>
-        /// Checks that ToJson() works properly for the custom JSON converter for type KeyboardShortcut.
-        /// </summary>
-        [Test]
-        [Category("JSON"), Category("Keyboard Shortcuts")]
-        public void KeyboardShortcutToJson()
-        {
-            JsonConverterSet converters = new JsonConverterSet(new KeyboardShortcutJsonConverter());
-
-            (KeyboardShortcut, JsonData)[] testCases =
-            {
-                (new KeyboardShortcut(CustomKeyCode.Ctrl, CustomKeyCode.Plus), new JsonList(new JsonString("Ctrl"), new JsonString("+"))),
-                (new KeyboardShortcut(KeyCode.G), new JsonList(new JsonString("G"))),
-                (KeyboardShortcut.None, new JsonList()),
-                (new KeyboardShortcut(KeyCode.Minus, CustomKeyCode._9, CustomKeyCode.Alt, CustomKeyCode.Shift),
-                new JsonList(new JsonString("Alt"), new JsonString("Shift"), new JsonString("9"), new JsonString("-"))),
-            };
-
-            foreach ((KeyboardShortcut keyboardShortcut, JsonData expected) in testCases)
-            {
-                Assert.True(JsonData.HaveSameData(JsonConverter.ToJson(keyboardShortcut, converters, false), expected));
-            }
-        }
-
-        /// <summary>
-        /// Checks that FromJson() works properly for the custom JSON converter for type KeyboardShortcut.
-        /// </summary>
-        [Test]
-        [Category("JSON"), Category("Keyboard Shortcuts")]
-        public void KeyboardShortcutFromJson()
-        {
-            JsonConverterSet converters = new JsonConverterSet(new KeyboardShortcutJsonConverter());
-
-            (KeyboardShortcut, JsonData)[] testCases =
-            {
-                (new KeyboardShortcut(CustomKeyCode.Ctrl, CustomKeyCode.Plus), new JsonList(new JsonString("Ctrl"), new JsonString("+"))),
-                (new KeyboardShortcut(KeyCode.G), new JsonList(new JsonString("G"))),
-                (KeyboardShortcut.None, new JsonList()),
-                (new KeyboardShortcut(KeyCode.Minus, CustomKeyCode._9, CustomKeyCode.Alt, CustomKeyCode.Shift),
-                new JsonList(new JsonString("Alt"), new JsonString("Shift"), new JsonString("9"), new JsonString("-"))),
-            };
-
-            foreach ((KeyboardShortcut expected, JsonData jsonData) in testCases)
-            {
-                Assert.AreEqual(JsonConverter.FromJson<KeyboardShortcut>(jsonData, converters, false), expected);
-            }
         }
     }
 }
