+++ Assets/Tests/KeyboardShortcutTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c1827e19391cc1c478c32a49c511c49c
